+++ R/RcppExports.R
@@ -57,6 +57,10 @@ getXPtrFBM <- function(path, n, m, type) {
     .Call(`_bigstatsr_getXPtrFBM`, path, n, m, type)
 }
 
+getXPtrFBM_RW <- function(path, n, m, type) {
+    .Call(`_bigstatsr_getXPtrFBM_RW`, path, n, m, type)
+}
+
 prod_FBM_mat <- function(BM, A) {
     .Call(`_bigstatsr_prod_FBM_mat`, BM, A)
 }