+++ tests/testthat/test-FBM-convert.R
@@ -42,131 +42,137 @@ expect_true(exists("XX"))
 
 ################################################################################
 
-get_text <- function(rtype, ctype) {
-  sprintf("while converting from R type '%s' to C type '%s'.", rtype, ctype)
-}
+test_that("Downcast warnings work", {
 
-for (gen in c("scalar", "vector", "matrix")) {
+  skip_on_cran()
 
-  # From double
-  x1 <- to_gen(runif(100))
-  # To raw
-  expect_warning(X <- FBM(10, 10, x1, type = "raw"),
-                 get_text("double", "unsigned char (raw)"), fixed = TRUE)
-  expect_warning(X[1:5] <- x1[1:5],
-                 get_text("double", "unsigned char (raw)"), fixed = TRUE)
-  expect_warning(X[2] <- NA_real_,
-                 get_text("double", "unsigned char (raw)"), fixed = TRUE)
-  expect_FBM(without_downcast_warning( FBM(10, 10, x1, type = "raw") ))
-  # To ushort
-  expect_warning(X <- FBM(10, 10, x1, type = "unsigned short"),
-                 get_text("double", "unsigned short"), fixed = TRUE)
-  expect_warning(X[1:5] <- x1[1:5],
-                 get_text("double", "unsigned short"), fixed = TRUE)
-  expect_warning(X[2] <- NA_real_,
-                 get_text("double", "unsigned short"), fixed = TRUE)
-  expect_FBM(without_downcast_warning( FBM(10, 10, x1, type = "unsigned short") ))
-  # To int
-  expect_warning(X <- FBM(10, 10, x1, type = "integer"),
-                 get_text("double", "integer"), fixed = TRUE)
-  expect_warning(X[1:5] <- x1[1:5],
-                 get_text("double", "integer"), fixed = TRUE)
-  X[2] <- NA_real_
-  expect_identical(X[2], NA_integer_)
-  expect_warning(X[2] <- Inf,
-                 get_text("double", "integer"), fixed = TRUE)
-  expect_warning(X[2] <- NaN,
-                 get_text("double", "integer"), fixed = TRUE)
-  expect_FBM(without_downcast_warning( FBM(10, 10, x1, type = "integer") ))
-  # To float
-  expect_warning(X <- FBM(10, 10, x1, type = "float"))
-  X[2] <- NA_real_
-  expect_identical(X[2], NA_real_)
-  # To double
-  expect_FBM(X <- FBM(10, 10, x1, type = "double"))
-  expect_identical(X[1:5] <- x1[1:5], x1[1:5])
-  X[2] <- NA_real_
-  expect_identical(X[2], NA_real_)
+  get_text <- function(rtype, ctype) {
+    sprintf("while converting from R type '%s' to C type '%s'.", rtype, ctype)
+  }
 
+  for (gen in c("scalar", "vector", "matrix")) {
 
-  # From integer
-  x2 <- to_gen(1:100 + 1e6L)
-  # To raw
-  expect_warning(X <- FBM(10, 10, x2, type = "raw"),
-                 get_text("integer", "unsigned char (raw)"), fixed = TRUE)
-  expect_warning(X[1:5] <- x2[1:5],
-                 get_text("integer", "unsigned char (raw)"), fixed = TRUE)
-  expect_warning(X[2] <- NA_integer_,
-                 get_text("integer", "unsigned char (raw)"), fixed = TRUE)
-  expect_FBM(without_downcast_warning( FBM(10, 10, x2, type = "raw") ))
-  # To ushort
-  expect_warning(X <- FBM(10, 10, x2, type = "unsigned short"),
-                 get_text("integer", "unsigned short"), fixed = TRUE)
-  expect_warning(X[1:5] <- x2[1:5],
-                 get_text("integer", "unsigned short"), fixed = TRUE)
-  expect_warning(X[2] <- NA_integer_,
-                 get_text("integer", "unsigned short"), fixed = TRUE)
-  expect_FBM(without_downcast_warning( FBM(10, 10, x2, type = "unsigned short") ))
-  # To int
-  expect_FBM(X <- FBM(10, 10, x2, type = "integer"))
-  expect_identical(X[1:5] <- x2[1:5], x2[1:5])
-  X[2] <- NA_integer_
-  expect_identical(X[2], NA_integer_)
-  # To float
-  expect_FBM(X <- FBM(10, 10, x2, type = "float"))
-  expect_identical(X[1:5] <- x2[1:5], x2[1:5])
-  X[2] <- NA_integer_
-  expect_identical(X[2], NA_real_)
-  # To double
-  expect_FBM(X <- FBM(10, 10, x2, type = "double"))
-  expect_identical(X[1:5] <- x2[1:5], x2[1:5])
-  X[2] <- NA_integer_
-  expect_identical(X[2], NA_real_)
+    # From double
+    x1 <- to_gen(runif(100))
+    # To raw
+    expect_warning(X <- FBM(10, 10, x1, type = "raw"),
+                   get_text("double", "unsigned char (raw)"), fixed = TRUE)
+    expect_warning(X[1:5] <- x1[1:5],
+                   get_text("double", "unsigned char (raw)"), fixed = TRUE)
+    expect_warning(X[2] <- NA_real_,
+                   get_text("double", "unsigned char (raw)"), fixed = TRUE)
+    expect_FBM(without_downcast_warning( FBM(10, 10, x1, type = "raw") ))
+    # To ushort
+    expect_warning(X <- FBM(10, 10, x1, type = "unsigned short"),
+                   get_text("double", "unsigned short"), fixed = TRUE)
+    expect_warning(X[1:5] <- x1[1:5],
+                   get_text("double", "unsigned short"), fixed = TRUE)
+    expect_warning(X[2] <- NA_real_,
+                   get_text("double", "unsigned short"), fixed = TRUE)
+    expect_FBM(without_downcast_warning( FBM(10, 10, x1, type = "unsigned short") ))
+    # To int
+    expect_warning(X <- FBM(10, 10, x1, type = "integer"),
+                   get_text("double", "integer"), fixed = TRUE)
+    expect_warning(X[1:5] <- x1[1:5],
+                   get_text("double", "integer"), fixed = TRUE)
+    X[2] <- NA_real_
+    expect_identical(X[2], NA_integer_)
+    expect_warning(X[2] <- Inf,
+                   get_text("double", "integer"), fixed = TRUE)
+    expect_warning(X[2] <- NaN,
+                   get_text("double", "integer"), fixed = TRUE)
+    expect_FBM(without_downcast_warning( FBM(10, 10, x1, type = "integer") ))
+    # To float
+    expect_warning(X <- FBM(10, 10, x1, type = "float"))
+    X[2] <- NA_real_
+    expect_identical(X[2], NA_real_)
+    # To double
+    expect_FBM(X <- FBM(10, 10, x1, type = "double"))
+    expect_identical(X[1:5] <- x1[1:5], x1[1:5])
+    X[2] <- NA_real_
+    expect_identical(X[2], NA_real_)
 
 
-  # From logical
-  x3 <- to_gen(sample(c(TRUE, FALSE), 100, TRUE))
-  # To raw
-  expect_FBM(X <- FBM(10, 10, x3, type = "raw"))
-  expect_identical(X[1:5] <- na.omit(x3[1:5]), na.omit(x3[1:5]))
-  expect_warning(X[2] <- NA,
-                 get_text("logical", "unsigned char (raw)"), fixed = TRUE)
-  # To ushort
-  expect_FBM(X <- FBM(10, 10, x3, type = "unsigned short"))
-  expect_identical(X[1:5] <- na.omit(x3[1:5]), na.omit(x3[1:5]))
-  expect_warning(X[2] <- NA,
-                 get_text("logical", "unsigned short"), fixed = TRUE)
-  # To int
-  expect_FBM(X <- FBM(10, 10, x3, type = "integer"))
-  expect_identical(X[1:5] <- x3[1:5], x3[1:5])
-  X[2] <- NA
-  expect_identical(X[2], NA_integer_)
-  # To float
-  expect_FBM(X <- FBM(10, 10, x3, type = "float"))
-  expect_identical(X[1:5] <- x3[1:5], x3[1:5])
-  X[2] <- NA
-  expect_identical(X[2], NA_real_)
-  # To double
-  expect_FBM(X <- FBM(10, 10, x3, type = "double"))
-  expect_identical(X[1:5] <- x3[1:5], x3[1:5])
-  X[2] <- NA
-  expect_identical(X[2], NA_real_)
+    # From integer
+    x2 <- to_gen(1:100 + 1e6L)
+    # To raw
+    expect_warning(X <- FBM(10, 10, x2, type = "raw"),
+                   get_text("integer", "unsigned char (raw)"), fixed = TRUE)
+    expect_warning(X[1:5] <- x2[1:5],
+                   get_text("integer", "unsigned char (raw)"), fixed = TRUE)
+    expect_warning(X[2] <- NA_integer_,
+                   get_text("integer", "unsigned char (raw)"), fixed = TRUE)
+    expect_FBM(without_downcast_warning( FBM(10, 10, x2, type = "raw") ))
+    # To ushort
+    expect_warning(X <- FBM(10, 10, x2, type = "unsigned short"),
+                   get_text("integer", "unsigned short"), fixed = TRUE)
+    expect_warning(X[1:5] <- x2[1:5],
+                   get_text("integer", "unsigned short"), fixed = TRUE)
+    expect_warning(X[2] <- NA_integer_,
+                   get_text("integer", "unsigned short"), fixed = TRUE)
+    expect_FBM(without_downcast_warning( FBM(10, 10, x2, type = "unsigned short") ))
+    # To int
+    expect_FBM(X <- FBM(10, 10, x2, type = "integer"))
+    expect_identical(X[1:5] <- x2[1:5], x2[1:5])
+    X[2] <- NA_integer_
+    expect_identical(X[2], NA_integer_)
+    # To float
+    expect_FBM(X <- FBM(10, 10, x2, type = "float"))
+    expect_identical(X[1:5] <- x2[1:5], x2[1:5])
+    X[2] <- NA_integer_
+    expect_identical(X[2], NA_real_)
+    # To double
+    expect_FBM(X <- FBM(10, 10, x2, type = "double"))
+    expect_identical(X[1:5] <- x2[1:5], x2[1:5])
+    X[2] <- NA_integer_
+    expect_identical(X[2], NA_real_)
 
 
-  # From raw
-  x4 <- to_gen(sample(as.raw(0:255), 100, TRUE))
-  expect_FBM(X <- FBM(10, 10, x4, type = "raw"))
-  expect_identical(X[1:5] <- x4[1:5], x4[1:5])
-  expect_FBM(X <- FBM(10, 10, x4, type = "unsigned short"))
-  expect_identical(X[1:5] <- x4[1:5], x4[1:5])
-  expect_FBM(X <- FBM(10, 10, x4, type = "integer"))
-  expect_identical(X[1:5] <- x4[1:5], x4[1:5])
-  expect_FBM(X <- FBM(10, 10, x4, type = "float"))
-  expect_identical(X[1:5] <- x4[1:5], x4[1:5])
-  expect_FBM(X <- FBM(10, 10, x4, type = "double"))
-  expect_identical(X[1:5] <- x4[1:5], x4[1:5])
+    # From logical
+    x3 <- to_gen(sample(c(TRUE, FALSE), 100, TRUE))
+    # To raw
+    expect_FBM(X <- FBM(10, 10, x3, type = "raw"))
+    expect_identical(X[1:5] <- na.omit(x3[1:5]), na.omit(x3[1:5]))
+    expect_warning(X[2] <- NA,
+                   get_text("logical", "unsigned char (raw)"), fixed = TRUE)
+    # To ushort
+    expect_FBM(X <- FBM(10, 10, x3, type = "unsigned short"))
+    expect_identical(X[1:5] <- na.omit(x3[1:5]), na.omit(x3[1:5]))
+    expect_warning(X[2] <- NA,
+                   get_text("logical", "unsigned short"), fixed = TRUE)
+    # To int
+    expect_FBM(X <- FBM(10, 10, x3, type = "integer"))
+    expect_identical(X[1:5] <- x3[1:5], x3[1:5])
+    X[2] <- NA
+    expect_identical(X[2], NA_integer_)
+    # To float
+    expect_FBM(X <- FBM(10, 10, x3, type = "float"))
+    expect_identical(X[1:5] <- x3[1:5], x3[1:5])
+    X[2] <- NA
+    expect_identical(X[2], NA_real_)
+    # To double
+    expect_FBM(X <- FBM(10, 10, x3, type = "double"))
+    expect_identical(X[1:5] <- x3[1:5], x3[1:5])
+    X[2] <- NA
+    expect_identical(X[2], NA_real_)
 
-}
+
+    # From raw
+    x4 <- to_gen(sample(as.raw(0:255), 100, TRUE))
+    expect_FBM(X <- FBM(10, 10, x4, type = "raw"))
+    expect_identical(X[1:5] <- x4[1:5], x4[1:5])
+    expect_FBM(X <- FBM(10, 10, x4, type = "unsigned short"))
+    expect_identical(X[1:5] <- x4[1:5], x4[1:5])
+    expect_FBM(X <- FBM(10, 10, x4, type = "integer"))
+    expect_identical(X[1:5] <- x4[1:5], x4[1:5])
+    expect_FBM(X <- FBM(10, 10, x4, type = "float"))
+    expect_identical(X[1:5] <- x4[1:5], x4[1:5])
+    expect_FBM(X <- FBM(10, 10, x4, type = "double"))
+    expect_identical(X[1:5] <- x4[1:5], x4[1:5])
+
+  }
+
+})
 
 ################################################################################